+++ pkg/deployments/tasks/sunset-utils/sense-v1-sunset-batch-1.json
@@ -1,0 +1,1 @@
+{"version":"1.0","chainId":"1","createdAt":"2023-10-17","meta":{"name":"sense-v1-sunset-batch-1","description":"","txBuilderVersion":"1.16.3","createdFromSafeAddress":"0xdd76360c26eaf63afcf3a8d2c0121f13ae864d57","createdFromOwnerAddress":"","checksum":"0x139b39157d208b191b96ff09710a3e74f78ab281de6e3e47cd3a7200b31ecb3f"},"transactions":[{"to":"0x93292717b1c0150a68a748c121be8c3b72dfefb8","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0xedd3b06b7596848e58e5e656e6b4973cd60be11a","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0xd28372e7b9904589d05bd257b85fea40fbd4df2b","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x9887e67aab4388ea4cf173b010df5c92b91f55b5","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0xa5874b6223c533be1b93aba1328a0582b1c6a863","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x13b72aefb9e7f0055fa5170f90f8da6b83dc23b1","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x304df20fcd8541bf750cf41b3d4b8551b9a3edc2","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x66e1ad7cda66a0b291aec63f3dbd8cb9eaf76680","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x60cd94a921c651723afcba8030951f2eb3ca572a","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x529c90e6d3a1aedab9b3011196c495439d23b893","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x8c5e7301a012dc677dd7dad97ae44032febcd0fd","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x0ed600eecf445b71ae3f8170cd368439e2739289","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x06ff1122636590c0797df2e0031060ac208e0e9f","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0xee15e6c4c6dacba162e6bd8c6c7185049beba212","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x86c55bffb64f8fcc8bea33f3176950fdd0fb160d","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x3c4d1e02abc4faa8d17b63ad897ba893cfdb0b37","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0xff37d3339f6aec38fbecab16b471e26d6607bdf8","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0xe59835238e63a413b93ba7b1fe65bed583266f97","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x2e1e993424cb2d3eb42c87d32bfd51b23854022d","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x5b487e3fcd7660d59f5af7cd6e56d6708a4b40ce","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x58b78221fe4a8aa5b02a4538a145a5c6da3afe0f","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0xad06efc15638bac0d9d5ba5497dfbe204956a53d","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null},{"to":"0x35c303e4be90422e03365b18ad59de02f7e85ff0","value":"0","data":"0x1a54259c0000000000000000000000000000000000000000000000000000000000000000","operation":"1","contractMethod":null,"contractInputsValues":null}]}
